requirements.txt
colorama==0.4.6
Flask==3.1.0
matplotlib==3.10.0
mediapipe==0.10.21
numpy==2.2.3
opencv_contrib_python==4.11.0.86
opencv_python==4.10.0.84
protobuf==5.29.3
python-dotenv==1.0.1
scikit_learn==1.6.1
tabulate==0.9.0
tqdm==4.67.1